Extract word suggestion to separate class
